api/requirements.txt
dbt-adapters==1.22.8
dbt-common==1.37.3
dbt-core==1.10.20
dbt-metricflow==0.11.0
dbt-postgres==1.8.0
dbt-protos==1.0.440
dbt-semantic-interfaces==0.9.4.dev0
fastapi==0.135.1
metricflow==0.209.0
psycopg2==2.9.11
psycopg2-binary==2.9.9
python-dotenv==1.2.2
uvicorn==0.41.0
